page routing, added building-en page, and updated room cards design on it
- moved dashboard component from app.js to a new file: dashboard.js
- updated app.js to handle routing for multiple pages
- added building-en.js for admin. view and search rooms in the EN building
- implemented new design layout for room display in en
